app/requirements.txt
fastapi
uvicorn
prometheus-client
cassandra-driver